pyproject.toml
[tool.poetry]
name = "kbol"
version = "0.1.0"
description = "Knowledge base from technical books using Ollama"
authors = ["Jason Walsh <j@wal.sh>"]
readme = "README.org"

[tool.poetry.dependencies]
python = "^3.11"
pypdf = "^3.17.1"
pgvector = "^0.2.3"
rich = "^13.7.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.25.2"
tiktoken = "^0.8.0"
files-to-prompt = "^0.4"
typer = "^0.9.0"
numpy = "^2.1.3"
black = "^24.10.0"
scikit-learn = "^1.4.0"
backoff = "^2.2.1"
prompt-toolkit = "^3.0.50"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kbol = "kbol.__main__:app"